API/requirements.txt
Flask
face_recognition
Pillow
numpy
